Доработаны примеры модуля 4.

+++ projects/examples/source/04.03.template.function.variadic.tree.traversal.cpp
@@ -23,7 +23,7 @@ int main()
 
 	for (auto i = 0uz; i < std::size(nodes); ++i)
 	{
-		nodes[i].data = static_cast < int > (i) + 1;
+		nodes[i].data = i + 1;
 	}
 
 	Node < int > * root = nullptr;
